Changed skip function to use long parameter

+++ variation-commons-mongodb/src/main/java/uk/ac/ebi/eva/commons/mongodb/services/VariantStudySummaryService.java
@@ -73,7 +73,7 @@ public class VariantStudySummaryService {
         Aggregation aggregation = Aggregation.newAggregation(
                 groupAndCount(),
                 projectAndFlatten(),
-                skip(pageNumber*pageSize),
+                skip((long)(pageNumber*pageSize)),
                 limit(pageSize)
         );
